Allow ordering JS/CSS resources.
When components rely on CSS or JS, sometimes they need to be included in
specific orders. Define a Directed Acyclic Graph of the JavaScript
resources and CSS resources we need, and render them in that order.
Allow callers to customize the rendering order by declaring a
relationship between two components. It is sufficient for either
component to know about the other; both declaring the same dependency
(i.e., A says it needs to be rendered before B, and B says it needs to
be rendered after A) is harmless but not necessary.

Also, make the linter happy, both by changing what we're linting on and
by adding comments just everywhere.

Finally, fix a bug where if a template tree is dependent on runtime data
(e.g., if templates are included/excluded based on the presence/absence
of specific data) then we'd not update the cache for the changing
template tree. Now we compare the files that make up the template tree
and the files we're trying to load before using a cached template.